WIP on walk code, removed gait library files. All code will be in walk state

+++ spot_micro_motion_cmd/include/spot_micro_motion_cmd/spot_micro_motion_cmd.h
@@ -49,7 +49,7 @@ struct SpotMicroNodeConfig {
   int swing_ticks;
   int stance_ticks;
   std::vector<int> phase_ticks;
-  float phase_length;
+  int phase_length;
 
 };
